moved levenstein distance function to a new base view class
